refactor: VideoHistoryFetcher 패키지 Video 하위로 이동 (#110)
