--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/domain/VideoHistoryFetcher.java
@@ -1,13 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.domain;
-
-import com.celuveat.celuveat.celeb.domain.Celeb;
-import com.celuveat.celuveat.video.domain.VideoHistory;
-import java.time.LocalDateTime;
-import java.util.List;
-
-public interface VideoHistoryFetcher {
-
-    List<VideoHistory> fetchAllVideoHistoriesByCeleb(Celeb celeb);
-
-    List<VideoHistory> fetchVideoHistoriesByCelebAfterDateTime(Celeb celeb, LocalDateTime startDateTime);
-}
