--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/exception/VideoHistoryFetcherExceptionType.java
@@ -1,35 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.exception;
-
-import com.celuveat.celuveat.common.exception.BaseExceptionType;
-import org.springframework.http.HttpStatus;
-
-public enum VideoHistoryFetcherExceptionType implements BaseExceptionType {
-
-    NOT_FOUND_RESTAURANT_LINK(400, HttpStatus.NOT_FOUND, "링크를 찾을 수 없습니다"),
-    ;
-
-    private final int errorCode;
-    private final HttpStatus httpStatus;
-    private final String errorMessage;
-
-    VideoHistoryFetcherExceptionType(int errorCode, HttpStatus httpStatus, String errorMessage) {
-        this.errorCode = errorCode;
-        this.httpStatus = httpStatus;
-        this.errorMessage = errorMessage;
-    }
-
-    @Override
-    public int errorCode() {
-        return errorCode;
-    }
-
-    @Override
-    public HttpStatus httpStatus() {
-        return httpStatus;
-    }
-
-    @Override
-    public String errorMessage() {
-        return errorMessage;
-    }
-}
