--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/infra/youtube/dto/search/Id.java
@@ -1,8 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.infra.youtube.dto.search;
-
-public record Id(
-        String kind,
-        String videoId,
-        String channelId
-) {
-}
