--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/infra/youtube/dto/search/PageInfo.java
@@ -1,7 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.infra.youtube.dto.search;
-
-public record PageInfo(
-        String totalResults,
-        String resultsPerPage
-) {
-}
