--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/infra/youtube/dto/search/Thumbnail.java
@@ -1,8 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.infra.youtube.dto.search;
-
-public record Thumbnail(
-        String url,
-        String width,
-        String height
-) {
-}
