--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/infra/youtube/dto/search/Thumbnails.java
@@ -1,13 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.infra.youtube.dto.search;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record Thumbnails(
-        @JsonProperty("default")
-        Thumbnail defaultThumbnail,
-        @JsonProperty("medium")
-        Thumbnail mediumThumbnail,
-        @JsonProperty("high")
-        Thumbnail highThumbnail
-) {
-}
