--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/infra/youtube/dto/video/Localized.java
@@ -1,7 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.infra.youtube.dto.video;
-
-public record Localized(
-        String title,
-        String description
-) {
-}
