--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/infra/youtube/dto/video/PageInfo.java
@@ -1,7 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.infra.youtube.dto.video;
-
-public record PageInfo(
-        int totalResults,
-        int resultsPerPage
-) {
-}
