--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/infra/youtube/dto/video/Statistics.java
@@ -1,9 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.infra.youtube.dto.video;
-
-public record Statistics(
-        String viewCount,
-        String likeCount,
-        String favoriteCount,
-        String commentCount
-) {
-}
