--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/infra/youtube/dto/video/Thumbnail.java
@@ -1,8 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.infra.youtube.dto.video;
-
-public record Thumbnail(
-        String url,
-        String width,
-        String height
-) {
-}
