--- backend/src/main/java/com/celuveat/celuveat/restaurantvideohistoryfetcher/infra/youtube/dto/video/Thumbnails.java
@@ -1,15 +1,0 @@
-package com.celuveat.celuveat.restaurantvideohistoryfetcher.infra.youtube.dto.video;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record Thumbnails(
-        @JsonProperty("default")
-        Thumbnail defaultThumbnail,
-        @JsonProperty("medium")
-        Thumbnail mediumThumbnail,
-        @JsonProperty("high")
-        Thumbnail highThumbnail,
-        @JsonProperty("standard")
-        Thumbnail standardThumbnail
-) {
-}
