+++ backend/src/main/java/com/celuveat/celuveat/video/domain/VideoHistoryFetcher.java
@@ -1,0 +1,12 @@
+package com.celuveat.celuveat.video.domain;
+
+import com.celuveat.celuveat.celeb.domain.Celeb;
+import java.time.LocalDateTime;
+import java.util.List;
+
+public interface VideoHistoryFetcher {
+
+    List<VideoHistory> fetchAllVideoHistoriesByCeleb(Celeb celeb);
+
+    List<VideoHistory> fetchVideoHistoriesByCelebAfterDateTime(Celeb celeb, LocalDateTime startDateTime);
+}
