+++ backend/src/main/java/com/celuveat/celuveat/video/exception/VideoHistoryFetcherException.java
@@ -1,0 +1,16 @@
+package com.celuveat.celuveat.video.exception;
+
+import com.celuveat.celuveat.common.exception.BaseException;
+import com.celuveat.celuveat.common.exception.BaseExceptionType;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class VideoHistoryFetcherException extends BaseException {
+
+    private final VideoHistoryFetcherExceptionType exceptionType;
+
+    @Override
+    public BaseExceptionType exceptionType() {
+        return exceptionType;
+    }
+}
