+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/search/Id.java
@@ -1,0 +1,8 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.search;
+
+public record Id(
+        String kind,
+        String videoId,
+        String channelId
+) {
+}
