+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/search/Item.java
@@ -1,0 +1,23 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.search;
+
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+
+public record Item(
+        String kind,
+        String etag,
+        Id id,
+        Snippet snippet
+) {
+
+    public LocalDateTime publishedAt() {
+        String publishedAt = snippet.publishedAt();
+        return LocalDateTime.from(
+                Instant.from(
+                        DateTimeFormatter.ISO_DATE_TIME.parse(publishedAt)
+                ).atZone(ZoneId.of("Asia/Seoul"))
+        );
+    }
+}
