+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/search/PageInfo.java
@@ -1,0 +1,7 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.search;
+
+public record PageInfo(
+        String totalResults,
+        String resultsPerPage
+) {
+}
