+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/search/SearchListResponse.java
@@ -1,0 +1,37 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.search;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public record SearchListResponse(
+        String kind,
+        String etag,
+        String nextPageToken,
+        String prevPageToken,
+        String regionCode,
+        PageInfo pageInfo,
+        List<Item> items
+) {
+
+    public List<String> videoIds() {
+        return items.stream()
+                .map(item -> item.id().videoId())
+                .collect(Collectors.toList());
+    }
+
+    public List<String> afterVideoIds(LocalDateTime startDateTime) {
+        return items.stream()
+                .filter(item -> item.publishedAt().isAfter(startDateTime))
+                .map(item -> item.id().videoId())
+                .collect(Collectors.toList());
+    }
+
+    public boolean isAllAfterVideo(List<String> videoIds) {
+        return items.size() == videoIds.size();
+    }
+
+    public boolean hasBeforeVideo(List<String> videoIds) {
+        return items.size() > videoIds.size();
+    }
+}
