+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/search/Snippet.java
@@ -1,0 +1,13 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.search;
+
+public record Snippet(
+        String publishedAt,
+        String channelId,
+        String title,
+        String description,
+        Thumbnails thumbnails,
+        String channelTitle,
+        String liveBroadcastContent,
+        String publishTime
+) {
+}
