+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/search/Thumbnail.java
@@ -1,0 +1,8 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.search;
+
+public record Thumbnail(
+        String url,
+        String width,
+        String height
+) {
+}
