+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/search/Thumbnails.java
@@ -1,0 +1,13 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.search;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record Thumbnails(
+        @JsonProperty("default")
+        Thumbnail defaultThumbnail,
+        @JsonProperty("medium")
+        Thumbnail mediumThumbnail,
+        @JsonProperty("high")
+        Thumbnail highThumbnail
+) {
+}
