+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/video/Localized.java
@@ -1,0 +1,7 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.video;
+
+public record Localized(
+        String title,
+        String description
+) {
+}
