+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/video/PageInfo.java
@@ -1,0 +1,7 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.video;
+
+public record PageInfo(
+        int totalResults,
+        int resultsPerPage
+) {
+}
