+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/video/Snippet.java
@@ -1,0 +1,19 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.video;
+
+import java.util.List;
+
+public record Snippet(
+        String publishedAt,
+        String channelId,
+        String title,
+        String description,
+        Thumbnails thumbnails,
+        String channelTitle,
+        List<String> tags,
+        String categoryId,
+        String liveBroadcastContent,
+        String defaultLanguage,
+        Localized localized,
+        String defaultAudioLanguage
+) {
+}
