+++ backend/src/main/java/com/celuveat/celuveat/video/infra/youtube/dto/video/Statistics.java
@@ -1,0 +1,9 @@
+package com.celuveat.celuveat.video.infra.youtube.dto.video;
+
+public record Statistics(
+        String viewCount,
+        String likeCount,
+        String favoriteCount,
+        String commentCount
+) {
+}
